Add SNWorks Support (#29)
* added ruby versioning and a new script to add ssh users

* Update Ruby and speed up docker-compose on mac

This updates Ruby to 2.5.7, and also speeds up booting on a Mac for
development purposes.

Affects #22, #21

* Remove mysql2 from requirements

* Update gems to newest versions

* Add Rubocop for linting

* Add Rubocop and begin linting cleanup

Because I've come to my senses, I now use Rubocop to lint all the code I
write in Ruby. This now includes this code base. I've enabled Rubocop
with a number of my own preferences, as well as include a Git hook to
automatically run everything on commit.

I also then began the clean up anew.

* Allow hypens(dashes) in host url names

Turns out that urls can have dashed '-' in them, this adds that
character to the check.

* Fix database issue with entrypoint.sh

* Fix docker local issues, move to webpack

This does a number of things:

- Update so the most modern version of docker runs
- Fix wp-paragraph issue with <p> tags for wordpress installs
- Moves to the modern Webpacker approach to handling js/css/etc

* Continuing work

* Adding files for merge

* Fix formatting for SEO Works

Fixes formatting for SEO Works, especially the following

- Proper date formatting
- If description is empy copy from body

* Add search functionality for SNWorks

This adds search API functionality for SNWorks.

* Fixes SSL cert folder

* Returns url in article for snworks

URL is not a standard response from the SNWorks API, using the format
{url}/article/{year}/{month, zero padded}/{slug} this returns the url
with the article json response for the app so sharing works.

* Adds Proper Push Notifications for SNWorks/FCM

FCM is the now-accepted version of Google push notifications, this
integrates that into a new version of Uniqush and deploys it for both
iOS and Android.

We also do the following:
- Add a headline option for all push notifications
- Display a dropdown when creating push notifications to choose an
  article instead of hunting for an id
- SNWorks fixes to add single article support
- Better error handling for creating a notification

* Forgot to add files

* Add RMagick gem for image dimension determination

* Add ImageMagick/RMagick for SEOWorks Image Sizes

SEOWorks seems to have a bug where occasionally images are returned
with valid URLs but with `null` for both the width and height dimensions.

Here we add ImageMagick and the RMagick gem for its Ruby bindings to
let us download the image and get hte dimentions ourselves if necessary.

If and when we hear back from SEOWorks this should be reverted/removed,
since it adds a huge amount of extra stuff that I'd prefer not to have
hanging around.

Rebuild all containers after upgrading

* updates ruby version

* Layout/Tabs is obsolete, replace

* Updates Yarn to 4

* Add yarn and node packs to non-nginx prod setup

We need a place to store webpacker etc. stuff, this adds volumes to do
so.

Rebuild containers after deployment

* Add ActiveJob to warm caches for articles

This affects all but is only tested for the moment on SNWorks.

This adds a job to warm the caches for articles. Since the round trip
can be very heavy (especially for SNWorks, where there are 20+ requests
per single api call), this caches the responses. For the moment every 30
seconds for testing.

* Migrate compose changes from dev to prod files

This moves all the changes we've been testing on the dev compose file
to the prod files so that redis and sidekiq get properly launched.

* Link sidekiq and redis to web

* Forgot to add scheduled_jobs to docker-compose

* Add sidekiq config file

* Add sidekiq to use the config file for docker compose

* Add sidekiq initializer

* Add sidekiq initializer

* add sidekiq to routes

* moved sidekiq initializer

* moved sidekiq initializer... back

* moved sidekiq initializer... back

* Move cache to 1 minute for testing

* comment out sidekiq

* Moved cache to Redis from memcache

* Forgot to change guard variable

* Adjusted caching times for production purposes

* Rearrange paid native advertising so its not top

The original organization that this integration was created for does
paid native advertising. The only way to determine which is paid however
is to look at the author, which in the original case is 'Scholarship
Media'. This will rearrange the articles so that the paid ad is never at
the top.

* Fix bug when handling paid advertisement

Forgot to break the loop when checking for a paid advertisement, causing
the entire array of articles to get processed. Whoops

+++ entrypoint.sh
@@ -1,4 +1,7 @@
-#!/bin/bash    
+#!/bin/bash
+echo ""
+echo "Running Bundler... 🎁"
+
 bundle config build.nokogiri --use-system-libraries
 bundle install --path vendor/bundle
 
@@ -16,8 +19,15 @@ cat >/etc/logrotate.conf <<EOL
 EOL
 
 # make sure the last run is cleared
-rm /push/tmp/pids/server.pid
+FILE=/push/tmp/pids/server.pid
+if test -f "$FILE"; then
+  echo "Previous PID file found @ $FILE"
+  echo "Deleting PID file..."
+  rm /push/tmp/pids/server.pid
+fi
 
+echo ""
+echo "🔺🔺🔺 Booting Rails in $RAILS_ENV mode 🔺🔺🔺"
 if [[ $RAILS_ENV = "development" ]]
 then
 	RAILS_ENV=development
@@ -28,18 +38,35 @@ then
   echo "Skipping precompile in test"
 else
   RAILS_ENV=production
-	echo "Precompiling assets..."
-	bundle exec rake assets:precompile 
-	echo "Done."
+	echo "Precompiling assets... 📦"
+	bundle exec rake assets:precompile
+  bundle exec rails webpacker:compile
+	echo "Done. 🎀"
 fi
 
-#if psql -lqt | cut -d \| -f 1 | grep -qw 'development'; then
-#    echo "database already exists"
-#else
-#	rake db:create
-#fi
+echo "Checking database status... 🔎"
+output="$(bundle exec rake db:abort_if_pending_migrations 2>&1 | grep -ci "ActiveRecord::NoDatabaseError")"
+if [[ $output = "1" ]]
+then
+  echo "Database not found, creating it now..."
+  bundle exec rake db:create
+  echo "Database created, moving onward 🎈"
+else
+  echo "Database exists, moving onward ➡️"
+fi
 
-#rake db:reset
-#rake db:migrate
+echo "Checking database migration status..."
+output="$(bundle exec rake db:abort_if_pending_migrations 2>&1 | grep -ci "update your database then try again.")"
+if [[ $output = "1" ]]
+then
+  echo "Database not up to date, running migrations..."
+  bundle exec rake db:migrate
+  echo "Database migrated, going forward 🎈"
+else
+  echo "Database up to date, going forward ➡️"
+fi
+
+echo "Database checks all done! 🔥"
+echo "If you're on a mac, please wait for awhile now... ⏳⏳⏳"
 
 exec $@
